requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
discord.py==2.5.2
Django==5.2
django-celery-beat==2.8.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
dotenv==0.9.9
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.30.2
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
textblob==0.19.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
